+++ common/src/main/java/com/iamkaf/amber/AmberMod.java
@@ -2,12 +2,15 @@ package com.iamkaf.amber;
 
 import com.iamkaf.amber.api.core.v2.AmberInitializer;
 import com.iamkaf.amber.api.core.v2.AmberModInfo;
+import com.iamkaf.amber.api.event.v1.events.common.ItemEvents;
+import com.iamkaf.amber.api.event.v1.events.common.PlayerEvents;
 import com.iamkaf.amber.api.platform.v1.Platform;
 import com.iamkaf.amber.command.AmberCommands;
 import com.iamkaf.amber.networking.v1.AmberNetworking;
 import com.iamkaf.amber.platform.Services;
 import com.iamkaf.amber.util.Env;
 import com.iamkaf.amber.util.EnvExecutor;
+import net.minecraft.world.InteractionResult;
 import org.jetbrains.annotations.Nullable;
 
 import java.util.ArrayList;
@@ -35,7 +38,6 @@ public class AmberMod {
         // Register commands
         AmberCommands.initialize();
     }
-
 
     /**
      * Gets the event bus for a specific Amber mod on Forge and NeoForge.